refactor: ESLint에 어긋나는 문법 삭제
- Homepage가 두번 임포트되어있어서 삭제했습니다.
- toRaw 사용이 없어서 삭제했습니다.
- GoogleChart의 이름이 임포트가 잘못되어있어 수정했습니다.

- issue: #29
